parent/django-etl/requirements.txt
Django>=2.0,<3.0
djangorestframework
py2neo
scikit-learn
